Remove layout component and move defaultroutes to the separate directory
