--- src/Layout.tsx
@@ -1,9 +1,0 @@
-import { Outlet } from 'react-router-dom';
-
-export function Layout() {
-  return (
-    <div>
-      <Outlet />
-    </div>
-  );
-}
